Remove compilation warnings in Calibration/HcalCalibAlgos and Calibration/HcalIsolatedTrackReco
