+++ Calibration/HcalIsolatedTrackReco/plugins/HITRegionalPixelSeedGenerator.cc
@@ -1,7 +1,6 @@
 //
 // Class:           HITRegionalPixelSeedGenerator
 
-#include "FWCore/Framework/interface/EDProducer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/ConsumesCollector.h"
 #include "DataFormats/Common/interface/Handle.h"
